GnewsWeb/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.7
cssselect==1.2.0
Django==4.2.5
django-jazzmin==2.6.0
dnspython==1.16.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.3
frozenlist==1.4.0
gnews==0.3.1
idna==3.4
jieba3k==0.35.1
joblib==1.3.2
lxml==4.9.3
multidict==6.0.4
newspaper3k==0.2.8
nltk==3.8.1
openai==0.28.0
Pillow==10.0.0
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==6.0.1
regex==2023.8.8
requests==2.26.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tinysegmenter==0.3
tldextract==3.5.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.16
yarl==1.9.2
